完成Android Embedding V2的适配，将SDK最低版本更为3.0.0，完成Flutter 3.13.x相关代码的适配

+++ example/lib/main.dart
@@ -20,9 +20,8 @@ void main() {
 }
 
 class MyApp extends StatefulWidget {
-
   const MyApp({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -31,13 +30,13 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   final cropKey = GlobalKey<CropState>();
-  File _file;
-  File _sample;
-  File _lastCropped;
-  ImageProvider decorator;
+  File? _file;
+  File? _sample;
+  File? _lastCropped;
+  ImageProvider? decorator;
 
-  ui.Image _decoratorImage;
-  ImageStream _decoratorImageStream;
+  ui.Image? _decoratorImage;
+  ImageStream? _decoratorImageStream;
 
   @override
   void initState() {
@@ -74,9 +73,8 @@ class _MyAppState extends State<MyApp> {
   }
 
   void test() async {
-    await Future.delayed(Duration(seconds: 1));
+    await Future.delayed(Duration(seconds: 1), () {});
     _getDecoratorImage();
-
   }
 
   @override
@@ -94,14 +92,14 @@ class _MyAppState extends State<MyApp> {
       children: <Widget>[
         Expanded(
           child: Crop.file(
-            _sample,
+            _sample!,
             key: cropKey,
             aspectRatio: 1,
             showGrid: false,
             enableAdjustCropWindow: false,
             onCalculateDefaultArea: _onCalculateDefaultArea,
             onAfterPaint: _onAfterPaint,
-            actionListener: (CropAction action){
+            actionListener: (CropAction action) {
               log("CropAction $action");
             },
           ),
@@ -117,8 +115,8 @@ class _MyAppState extends State<MyApp> {
                   'Crop Image',
                   style: Theme.of(context)
                       .textTheme
-                      .button
-                      .copyWith(color: Colors.white),
+                      .labelLarge
+                      ?.copyWith(color: Colors.white),
                 ),
                 onPressed: () => _cropImage(),
               ),
@@ -146,7 +144,7 @@ class _MyAppState extends State<MyApp> {
     if (_decoratorImage == null) {
       return;
     }
-    ui.Image image = _decoratorImage;
+    ui.Image image = _decoratorImage!;
     canvas.drawImageRect(
         image,
         Rect.fromLTWH(0, 0, image.width.toDouble(), image.height.toDouble()),
@@ -158,7 +156,10 @@ class _MyAppState extends State<MyApp> {
     return TextButton(
       child: Text(
         'Open Image',
-        style: Theme.of(context).textTheme.button.copyWith(color: Colors.white),
+        style: Theme.of(context)
+            .textTheme
+            .labelLarge
+            ?.copyWith(color: Colors.white),
       ),
       onPressed: () => _openImage(),
     );
@@ -167,10 +168,11 @@ class _MyAppState extends State<MyApp> {
   Future<void> _openImage() async {
     final pickedFile =
         await ImagePicker().getImage(source: ImageSource.gallery);
+    if (pickedFile == null) return;
     final file = File(pickedFile.path);
     final sample = await ImageCrop.sampleImage(
       file: file,
-      preferredSize: context.size.longestSide.ceil(),
+      preferredSize: context.size?.longestSide.ceil(),
     );
 
     _sample?.delete();
@@ -183,9 +185,9 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> _cropImage() async {
-    final scale = cropKey.currentState.scale;
-    final area = cropKey.currentState.area;
-    if (area == null) {
+    final scale = cropKey.currentState?.scale;
+    final area = cropKey.currentState?.area;
+    if (scale == null || area == null || _file == null) {
       // cannot crop, widget is not setup
       return;
     }
@@ -193,7 +195,7 @@ class _MyAppState extends State<MyApp> {
     // scale up to use maximum possible number of pixels
     // this will sample image in higher resolution to make cropped image larger
     final sample = await ImageCrop.sampleImage(
-      file: _file,
+      file: _file!,
       preferredSize: (2000 / scale).round(),
     );
 
@@ -215,11 +217,11 @@ class _MyAppState extends State<MyApp> {
 
     final oldImageStream = _decoratorImageStream;
     _decoratorImageStream =
-        decorator.resolve(createLocalImageConfiguration(context));
-    if (_decoratorImageStream.key != oldImageStream?.key) {
+        decorator!.resolve(createLocalImageConfiguration(context));
+    if (_decoratorImageStream!.key != oldImageStream?.key) {
       oldImageStream
           ?.removeListener(ImageStreamListener(_updateDecoratorImage));
-      _decoratorImageStream
+      _decoratorImageStream!
           .addListener(ImageStreamListener(_updateDecoratorImage));
     }
   }
